Convert "unknown version" versions
The "unknown version" is used as the target when a feature has been
supported by a browser for a long time, but no one is sure when support
was added, and it was probably supported in the first released version.
This was encoded as version=null, status=unknown, but is now encoded as
version="current", status="current".